requirements.txt
Django==5.1
djangorestframework==3.15.2
mysqlclient==2.2.4
requests==2.31.0
python-dotenv==1.0.1
django-autoslug==1.9.9
python-jose==3.3.0
drf-yasg==1.21.7